+++ src/extension.ts
@@ -21,6 +21,7 @@ import { OverlayMenu } from "./features/overlayMenu.js"
 import { MenuAnimation } from "./features/menuAnimation.js"
 import { DebugFeature } from "./features/debug.js"
 import { VolumeMixerWidgetFeature } from "./features/widget/volumeMixer.js"
+import { SystemIndicatorLayoutFeature } from "./features/layout/systemIndicator.js"
 
 export default class QstExtension extends Extension {
 	private features: FeatureBase[]
@@ -65,6 +66,7 @@ export default class QstExtension extends Extension {
 			new MenuAnimation(),
 			new SystemItemsLayoutFeature(),
 			new TogglesLayoutFeature(),
+			new SystemIndicatorLayoutFeature(),
 		]
 
 		// Load debug feature