+++ src/features/menuAnimation.ts
@@ -84,7 +84,7 @@ export class MenuAnimation extends FeatureBase {
 		}
 
 		this.tracker = new QuickSettingsMenuTracker()
-		this.tracker.onOpen = this.onOpen.bind(this)
+		this.tracker.onMenuOpen = this.onOpen.bind(this)
 		this.tracker.load()
 	}
 	override onUnload(): void {