+++ src/libs/shell/quickSettingsUtils.ts
@@ -1,22 +1,27 @@
 import {
 	QuickMenuToggle,
 	QuickToggle,
+	SystemIndicator,
 	type QuickSettingsMenu,
 } from "resource:///org/gnome/shell/ui/quickSettings.js"
 import { type PopupMenuBase, PopupSeparatorMenuItem } from "resource:///org/gnome/shell/ui/popupMenu.js"
 import { Global } from "../../global.js"
 import Maid from "../shared/maid.js"
 
-export abstract class QuickSettingsTrackerBase<T> {
+export abstract class ChildrenTrackerBase<T> {
 	appliedChild: Map<T, Maid>
-	gridConnection: number
+	addConnection: number
+	connectTarget: any
+	protected abstract getConnectTarget(): any
+	protected abstract catchChild(child: any): void
 	load(): void {
+		const connectTarget = this.connectTarget = this.getConnectTarget()
 		this.appliedChild = new Map()
-		this.gridConnection = Global.QuickSettingsGrid.connect("child-added", (_: any, child: any)=>{
-			const ok = this.catchChild(child)
-			if (ok && this.onUpdate) this.onUpdate()
+		this.addConnection = connectTarget.connect("child-added", (_: any, child: any)=>{
+			this.catchChild(child)
+			if (this.onUpdate) this.onUpdate()
 		})
-		for (const child of Global.QuickSettingsGrid.get_children()) {
+		for (const child of connectTarget.get_children()) {
 			this.catchChild(child)
 		}
 		if (this.onUpdate) this.onUpdate()
@@ -25,39 +30,40 @@ export abstract class QuickSettingsTrackerBase<T> {
 		for (const maid of this.appliedChild.values()) {
 			maid.destroy()
 		}
-		Global.QuickSettingsGrid.disconnect(this.gridConnection)
-		this.gridConnection = null
+		this.connectTarget.disconnect(this.addConnection)
+		this.addConnection = null
 		this.appliedChild = null
 	}
 	get items(): T[] {
 		if (!this.appliedChild) return []
 		return [...this.appliedChild.keys()]
 	}
-	protected abstract catchChild(child: any): boolean
 	onUpdate: ()=>void
 }
 
-export class QuickSettingsMenuTracker extends QuickSettingsTrackerBase<QuickSettingsMenu> {
-	onOpen: (maid: Maid, menu: QuickSettingsMenu, isOpen: boolean)=>void
+export class QuickSettingsMenuTracker extends ChildrenTrackerBase<QuickSettingsMenu> {
+	onMenuOpen: (maid: Maid, menu: QuickSettingsMenu, isOpen: boolean)=>void
 	onMenuCreated: (maid: Maid, menu: QuickSettingsMenu)=>void
-	protected override catchChild(child: any): boolean {
+	protected override catchChild(child: any): void {
 		const menu = child.menu
-		if (!menu) return false
-		if (this.appliedChild.has(menu)) return false
+		if (!menu) return
+		if (this.appliedChild.has(menu)) return
 
 		const menuMaid = new Maid()
 		menuMaid.functionJob(()=>{
 			this.appliedChild.delete(menu)
 		})
 		menuMaid.connectJob(menu, "open-state-changed", (_: any, isOpen: boolean) => {
-			if (this.onOpen) this.onOpen(menuMaid, menu, isOpen)
+			if (this.onMenuOpen) this.onMenuOpen(menuMaid, menu, isOpen)
 		})
 		menuMaid.connectJob(menu, "destroy", ()=>{
 			menuMaid.destroy()
 		})
 		if (this.onMenuCreated) this.onMenuCreated(menuMaid, menu)
 		this.appliedChild.set(menu, menuMaid)
-		return true
+	}
+	protected override getConnectTarget() {
+		return Global.QuickSettingsGrid
 	}
 	get menus() {
 		if (!this.appliedChild) return []
@@ -65,14 +71,14 @@ export class QuickSettingsMenuTracker extends QuickSettingsTrackerBase<QuickSett
 	}
 }
 
-export class QuickSettingsToggleTracker extends QuickSettingsTrackerBase<QuickToggle|QuickMenuToggle> {
+export class QuickSettingsToggleTracker extends ChildrenTrackerBase<QuickToggle|QuickMenuToggle> {
 	onToggleCreated: (maid: Maid, toggle: QuickToggle|QuickMenuToggle)=>void
-	protected override catchChild(child: any): boolean {
+	protected override catchChild(child: any): void {
 		if (
 			!(child instanceof QuickToggle)
 			&& !(child instanceof QuickMenuToggle)
-		) return false
-		if (this.appliedChild.has(child)) return false
+		) return
+		if (this.appliedChild.has(child)) return
 
 		const toggleMaid = new Maid()
 		toggleMaid.functionJob(()=>{
@@ -83,7 +89,32 @@ export class QuickSettingsToggleTracker extends QuickSettingsTrackerBase<QuickTo
 		})
 		if (this.onToggleCreated) this.onToggleCreated(toggleMaid, child)
 		this.appliedChild.set(child, toggleMaid)
-		return true
+	}
+	protected override getConnectTarget() {
+		return Global.QuickSettingsGrid
+	}
+}
+
+export class SystemIndicatorTracker extends ChildrenTrackerBase<SystemIndicator> {
+	onIndicatorCreated: (maid: Maid, indicator: SystemIndicator)=>void
+	protected override catchChild(child: any): void {
+		if (
+			!(child instanceof SystemIndicator)
+		) return
+		if (this.appliedChild.has(child)) return
+
+		const indicatorMaid = new Maid()
+		indicatorMaid.functionJob(()=>{
+			this.appliedChild.delete(child)
+		})
+		indicatorMaid.connectJob(child, "destroy", ()=>{
+			indicatorMaid.destroy()
+		})
+		if (this.onIndicatorCreated) this.onIndicatorCreated(indicatorMaid, child)
+		this.appliedChild.set(child, indicatorMaid)
+	}
+	protected override getConnectTarget() {
+		return Global.Indicators
 	}
 }